fix: SubjectErasureOutcome.Erased is int64 to match purger row counts
The purger APIs return int64 row counts (DeleteSagasBySubject / DeleteOutboxBySubject
/ DeleteIdempotencyBySubject, ultimately sql.Result.RowsAffected). Erased was int, so
those counts were narrowed via int(n) — which truncates on 32-bit builds and is the
wrong type for a row count on any platform, risking incorrect accountability figures
in ErasureResult and the erasure certificate. Make Erased int64 and drop the casts;
the snapshot eraser's local counter follows suit.

+++ adapters/postgres/subject_index_test.go
@@ -106,7 +106,7 @@ func TestSnapshotSubjectEraser_Postgres(t *testing.T) {
 	out, err := mink.NewSnapshotSubjectEraser(adapter).EraseSubject(ctx, "u1", fp)
 	require.NoError(t, err)
 	assert.Equal(t, "snapshot", out.Name)
-	assert.Equal(t, 1, out.Erased)
+	assert.Equal(t, int64(1), out.Erased)
 
 	snap, err := adapter.LoadSnapshot(ctx, "User-u1")
 	require.NoError(t, err)
